Style the markdown: Factor out boilerplate
