+++ context/app/help/help_app.py
@@ -1,8 +1,10 @@
 import dash
 import dash_html_components as html
+import dash_core_components as dcc
 
+from app.resource_loader import ResourceLoader
 
-class HelpApp():
+class HelpApp(ResourceLoader):
 
     def __init__(self,
                  server=None,
@@ -10,6 +12,8 @@ class HelpApp():
                  api_prefix=None):
         self.app = dash.Dash(server=server,
                              url_base_pathname=url_base_pathname)
+        self.load_resources()
+
         if api_prefix:
             self.app.config.update({
                 'requests_pathname_prefix': api_prefix
@@ -18,8 +22,12 @@ class HelpApp():
 
         self.app.layout = html.Div([
             html.Div([
-                html.Div([
-                    'help???'
-                ], className='col-md-6'),
+                dcc.Markdown([
+'''
+# Help
+
+This is a test
+'''
+                ], className='col-md-12'),
             ], className='row'),
         ], className='container')
