+++ context/app/vis/base.py
@@ -8,8 +8,7 @@ from plotly.figure_factory.utils import PLOTLY_SCALES
 from app.utils.pca import pca
 from app.utils.search import SimpleIndex
 
-
-class VisBase:
+class VisBase():
 
     def __init__(self,
                  union_dataframe,
@@ -38,21 +37,6 @@ class VisBase:
             self._color_scale = list(reversed(PLOTLY_SCALES[colors]))
         else:
             self._color_scale = PLOTLY_SCALES[colors]
-        with open(relative_path('../extra.css')) as extra_css_file:
-            self._css_urls = [
-                'https://maxcdn.bootstrapcdn.com/'
-                'bootstrap/3.3.7/css/bootstrap.min.css',
-                to_data_uri(extra_css_file.read(), 'text/css')
-            ]
-        with open(relative_path('../extra.js')) as extra_js_file:
-            self._js_urls = [
-                'https://code.jquery.com/'
-                'jquery-3.1.1.slim.min.js',
-                'https://maxcdn.bootstrapcdn.com/'
-                'bootstrap/3.3.7/js/bootstrap.min.js',
-                to_data_uri(extra_js_file.read(),
-                            'application/javascript')
-            ]
         self._debug = debug
         self.app = dash.Dash(server=server,
                              url_base_pathname=url_base_pathname)
@@ -69,16 +53,3 @@ class VisBase:
     #     if self._debug:
     #         # TODO: logging.info() didn't work. Check logging levels?
     #         print(' | '.join([str(field) for field in fields]))
-
-
-def relative_path(file):
-    # https://stackoverflow.com/questions/4060221 for more options
-    return os.path.join(os.path.dirname(__file__), file)
-
-
-def to_data_uri(s, mime):
-    uri = (
-        ('data:' + mime + ';base64,').encode('utf8') +
-        urlsafe_b64encode(s.encode('utf8'))
-    ).decode("utf-8", "strict")
-    return uri
